service discovery: move `GetServiceName` into own package
