+++ pkg/collector/corechecks/servicediscovery/detector/service_detector.go
@@ -3,7 +3,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2016-present Datadog, Inc.
 
-package servicediscovery
+// Package detector provides service metadata detection.
+package detector
 
 import (
 	"slices"
